tests/krate/publish: Extract `audit_action` module
